origin check on transferFrom
The vast majority of use legitimate cases for transferFrom are for sending tokens to an approved smart contract in a transaction initiated and signed by the token holder. These are generally the intent of the approver. The ability for tokens to be transferred in a transaction that is not initiated by the token holder is generally not the intent of the approver, and can easily be exploited by e.g. a proxy contract.